refactor(mcp): extract pure helpers to utils/ (http, mermaid, nodeId)
Phase 4 of McpServer.ts decomposition:
- utils/http.ts: cors, sendJson, replyError, collectBody, isValidHost
- utils/mermaid.ts: parseMermaid flowchart parser
- utils/nodeId.ts: normalizeNodeId, collapsing 6 inline duplicates

All pure functions, no class-state dependency. No behavior change.
McpServer.ts now ~978 lines (from 2411).
